requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
argon2-cffi==23.1.0
PyJWT==2.9.0
pydantic-settings==2.6.1
python-json-logger==2.0.7
user-agents==2.2.0
tldextract==5.1.2
